puse para convertir un int a un double y estaba mal, por lo que lo cambie

+++ programa1/programa1/Comandas.cs
@@ -296,7 +296,7 @@ namespace programa1
                     if (datos.Read())
                     {
                         precio = Convert.ToDouble(datos["precio"]);
-                        fila.Cells["Columna_Subtotal"].Value = Convert.ToDouble(fila.Cells["Columna_Cantidad"].Value) * precio;
+                        fila.Cells["Columna_Subtotal"].Value = Convert.ToInt32(fila.Cells["Columna_Cantidad"].Value) * precio;
                         acumulador = acumulador + (Convert.ToInt32(fila.Cells["Columna_Cantidad"].Value) * precio);
                     }
                     datos.Close();
